add new base class
