+++ src/ragas/testset/synthesizers/__init__.py
@@ -5,7 +5,9 @@ from ragas.testset.synthesizers.multi_hop import (
     MultiHopAbstractQuerySynthesizer,
     MultiHopSpecificQuerySynthesizer,
 )
-from ragas.testset.synthesizers.single_hop import SingleHopQuerySynthesizer
+from ragas.testset.synthesizers.single_hop.specific import (
+    SingleHopSpecificQuerySynthesizer,
+)
 
 from .base import BaseSynthesizer
 
@@ -15,7 +17,7 @@ QueryDistribution = t.List[t.Tuple[BaseSynthesizer, float]]
 def default_query_distribution(llm: BaseRagasLLM) -> QueryDistribution:
     """ """
     return [
-        (SingleHopQuerySynthesizer(llm=llm), 0.5),
+        (SingleHopSpecificQuerySynthesizer(llm=llm), 0.5),
         (MultiHopAbstractQuerySynthesizer(llm=llm), 0.25),
         (MultiHopSpecificQuerySynthesizer(llm=llm), 0.25),
     ]
